Move lint report pipeline out of cmd
